pymongo

+++ channel/views.py
@@ -7,7 +7,7 @@ from main_app.models import Employee
 from .models import channel_model
 from django.contrib.auth.models import User
 from video.models import video_class
-
+from main_app.views import logger_history_function
 
 
 @api_view(['POST'])
@@ -20,9 +20,10 @@ def create_channel(request):
         title = request.POST.get('title')
         description = request.POST.get('description')
         if channel_model.objects.filter(title=title):
-            message = 'channel name already register by same name'
+            message = 'channel name already register by same name '+title
             error = 'True'
             data = {'message ': message, 'error': error}
+            logger_history_function(token,message)
         elif Token.objects.filter(key=token):
             token_obj = Token.objects.get(key=token)
             user = token_obj.user
@@ -34,14 +35,16 @@ def create_channel(request):
                 employee.channel_id = channel_obj.id
                 print(channel_obj.id)
                 employee.save()
-                message = 'channel create successfull'
+                message = 'channel create successfully channel name '+title
                 error = 'False'
                 data = {'message ': message, 'error': error}
+                logger_history_function(token, message)
             else:  # if channel already exist for same user
                 channel = channel_model.objects.get(id=employee.channel_id)
                 message = 'channel already exist with channel name : ' + channel.title
-                error = 'true'
+                error = 'True'
                 data = {'message ': message, 'error': error}
+                logger_history_function(token, message)
 
         else:  # if  Token not found in database means user not exit
             message = 'token not found'
@@ -65,26 +68,36 @@ def delete_channel(request):
                 message = 'This User has no channel exit'
                 error = 'False'
                 data = {'message ': message, 'error': error}
+                logger_history_function(token, message)
             else:  # if channel already exist for same user
-                channel = channel_model.objects.get(id=channel_id)
-                video_ids =channel.video_id
-                print(video_ids)
-                video_list = video_ids.split(",")
-                print(video_list)
-                video_list = [s for s in video_list if s.isdigit()]
-                for id in video_list :
-                    print(id)
-                    if video_class.objects.filter(id=id) :
-                        video_obj = video_class.objects.get(id=id)
-                        video_obj.delete()
-                    else:
-                        print(f'{id},video not found')
-                employee.channel_id=0
-                employee.save()
-                channel.delete()
-                message = 'channel delete successfull'
-                error = 'False'
-                data = {'message ': message, 'error': error}
+                if not channel_model.objects.filter(id=channel_id):
+                    message = 'cannot delete channel channel ids wrong '
+                    error = 'True'
+                    data = {'message ': message, 'error': error}
+                    logger_history_function(token, message)
+
+                else:
+                    channel = channel_model.objects.get(id=channel_id)
+                    title = channel.title
+                    video_ids =channel.video_id
+                    print(video_ids)
+                    video_list = video_ids.split(",")
+                    print(video_list)
+                    video_list = [s for s in video_list if s.isdigit()]
+                    for id in video_list :
+                        print(id)
+                        if video_class.objects.filter(id=id) :
+                            video_obj = video_class.objects.get(id=id)
+                            video_obj.delete()
+                        else:
+                            print(f'{id},video not found')
+                    employee.channel_id=0
+                    employee.save()
+                    channel.delete()
+                    message = 'channel delete successfully : '+title
+                    error = 'False'
+                    data = {'message ': message, 'error': error}
+                    logger_history_function(token, message)
 
         else:  # if  Token not found in database means user not exit
             message = 'token not found'
